bump 0.2.0

+++ lib/src/convert/discount.dart
@@ -18,5 +18,5 @@ num discountToBackend(num value) {
       ? value
       : double.parse(truncateNumber(value, decimals: 1)),
     1000
-  );
+  ).toInt();
 }
