+++ lib/src/convert/money.dart
@@ -9,5 +9,5 @@ num moneyToDisplay(int value, {int unit = MONEY_YUAN_TO_CENT}) {
 
 /// 把金额转换为后端接口所用的格式
 num moneyToBackend(num value, {int unit = MONEY_YUAN_TO_CENT}) {
-  return timesNumber(value, unit);
+  return timesNumber(value, unit).toInt();
 }
